pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-return-desk-env"
version = "0.1.0"
description = "ReturnDesk OpenEnv environment for e-commerce returns and refunds"
requires-python = ">=3.10"
dependencies = [
  "openenv-core[core]>=0.2.2",
  "fastapi>=0.115.0",
  "pydantic>=2.0.0",
  "uvicorn[standard]>=0.24.0",
  "requests>=2.31.0",
  "openai>=1.0.0",
  "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-cov>=4.0.0",
]

[project.scripts]
server = "return_desk_env.server.app:main"

[tool.setuptools]
packages = ["return_desk_env", "return_desk_env.server", "return_desk_env.tasks"]
package-dir = { "return_desk_env" = ".", "return_desk_env.server" = "server", "return_desk_env.tasks" = "tasks" }

[tool.setuptools.package-data]
return_desk_env = ["*.yaml", "*.yml", "README.md"]
